Commit related to issue #10
Issue related possibly resolved needs to testing, but the change is very minor and should just work.

Updated Unity version; Should not cause issues for things, but if it does we can revert simply enough.

Also fixed the mod project file to use environment variable HOLLOW_KNIGHT_ROOT for game locations in the future. Been lazy on this one, and this is a good time to fix it.